Graphical presentation
We solved the problems related to the equations already established. We are facing one problem at the moment that we are working on, it actually concerns the nose of the robot. We are dealing with 2 second order equations with two unknown that gives us sometimes solutions in C instead of R. More uploads will come once we tackle this ad make more progress.